Day 69 Update 2
Added most functionality to the site.

Made a redirect when trying to access content that needs user to be signed in, sends to login page, and then returns to original destination if possible.

Security change that limits the post editing and deletion to the person who originally posted the post.
 - Will/should be adding a bypass for admins, so they can freely edit others' posts and delete if needed.

+++ Day 69 - Blog Capstone - Adding Users/main.py
@@ -1,8 +1,10 @@
+import flask
 from flask import Flask, render_template, redirect, url_for, flash
 from flask_bootstrap import Bootstrap
 from flask_ckeditor import CKEditor
 from datetime import date
 from werkzeug.security import generate_password_hash, check_password_hash
+from werkzeug.urls import url_parse
 from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy.orm import relationship
 from flask_login import UserMixin, login_user, LoginManager, login_required, current_user, logout_user
@@ -22,6 +24,20 @@ db = SQLAlchemy(app)
 ##Setup Flask Login
 login_manager = LoginManager()
 login_manager.init_app(app)
+login_manager.login_view = 'login'
+
+
+##CONFIGURE TABLES
+class BlogPost(db.Model):
+    __tablename__ = "blog_posts"
+    id = db.Column(db.Integer, primary_key=True)
+    author = db.Column(db.String(250), nullable=False)
+    author_account = db.Column(db.String(100), nullable=False)
+    title = db.Column(db.String(250), unique=True, nullable=False)
+    subtitle = db.Column(db.String(250), nullable=False)
+    date = db.Column(db.String(250), nullable=False)
+    body = db.Column(db.Text, nullable=False)
+    img_url = db.Column(db.String(250), nullable=False)
 
 
 class User(UserMixin, db.Model):
@@ -30,6 +46,12 @@ class User(UserMixin, db.Model):
     email = db.Column(db.String(100), unique=True)
     password = db.Column(db.String(100))
     name = db.Column(db.String(100))
+
+    def set_password(self, password):
+        self.password = generate_password_hash(password, method="pbkdf2:sha256", salt_length=8)
+
+    def check_password(self, password):
+        return check_password_hash(self.password, password)
 
 
 @login_manager.user_loader
@@ -41,55 +63,94 @@ def is_logged_in():
     return current_user.is_authenticated
 
 
-##CONFIGURE TABLES
-
-class BlogPost(db.Model):
-    __tablename__ = "blog_posts"
-    id = db.Column(db.Integer, primary_key=True)
-    author = db.Column(db.String(250), nullable=False)
-    title = db.Column(db.String(250), unique=True, nullable=False)
-    subtitle = db.Column(db.String(250), nullable=False)
-    date = db.Column(db.String(250), nullable=False)
-    body = db.Column(db.Text, nullable=False)
-    img_url = db.Column(db.String(250), nullable=False)
-
-
 @app.route('/')
 def get_all_posts():
 
     posts = db.session.query(BlogPost).all()
-
     return render_template("index.html", all_posts=posts)
 
 
-@app.route('/register')
+def login_failed():
+    flash("Either this email doesn't exist, or the password entered was incorrect.")
+
+
+def check_for_existing_user(email):
+    """
+        Returns true if email is already in database, false if not.
+    """
+    user = db.session.query(User).filter_by(email=email).first()
+    return user is not None
+
+
+def get_next_page():
+    next_page = flask.request.args.get('next')
+    if not next_page or url_parse(next_page).netloc != '':
+        next_page = url_for('get_all_posts')
+
+    return next_page
+
+
+@app.route('/register', methods=['GET', 'POST'])
 def register():
+
+    # On the off chance the user typed in the url to register:
+    # Check to make sure user isn't already signed in
+    if is_logged_in():
+        return redirect(url_for('get_all_posts'))
 
     form = RegisterNewUserForm()
     if form.validate_on_submit():
-        # @TODO add new user register logic
-        pass
+
+        if not check_for_existing_user(form.email.data):
+            new_user = User(
+                email=form.email.data,
+                name=form.name.data
+            )
+            new_user.set_password(form.password.data)
+
+            db.session.add(new_user)
+            db.session.commit()
+            db.session.refresh(new_user)
+
+            login_user(new_user)
+
+            return redirect(get_next_page())
+        else:
+            flash('This email is already in use.')
 
     return render_template("register.html", form=form)
 
 
-@app.route('/login')
+@app.route('/login', methods=['GET', 'POST'])
 def login():
 
     form = LoginForm()
     if form.validate_on_submit():
-        # @TODO add logging in logic
-        pass
+
+        user = db.session.query(User).filter_by(email=form.email.data).first()
+        if user is not None:
+
+            if user.check_password(form.password.data):
+                login_user(user)
+
+                return redirect(get_next_page())
+            else:
+                login_failed()
+        else:
+            login_failed()
 
     return render_template("login.html", form=form)
 
 
 @app.route('/logout')
 def logout():
-    if is_logged_in():
+
+    # If we're not logged in, redirect to main page.
+    # Wouldn't make sense to logout when already logged out.
+    if not is_logged_in():
         return redirect(url_for('get_all_posts'))
 
-    # @TODO add the rest of the logging out logic before going to main screen
+    logout_user()
 
     return redirect(url_for('get_all_posts'))
 
@@ -114,7 +175,8 @@ def contact():
     return render_template("contact.html")
 
 
-@app.route("/new-post")
+@app.route("/new-post", methods=['GET', 'POST'])
+@login_required
 def add_new_post():
 
     form = CreatePostForm()
@@ -125,7 +187,8 @@ def add_new_post():
             subtitle=form.subtitle.data,
             body=form.body.data,
             img_url=form.img_url.data,
-            author=current_user,
+            author=current_user.name,
+            author_account=current_user.email,
             date=date.today().strftime("%B %d, %Y")
         )
 
@@ -137,10 +200,20 @@ def add_new_post():
     return render_template("make-post.html", form=form)
 
 
+def check_permissions(post):
+    return current_user.email == post.author_account
+
+
 @app.route("/edit-post/<int:post_id>")
+@login_required
 def edit_post(post_id):
 
     post = db.session.query(BlogPost).get(post_id)
+
+    if not check_permissions(post):
+        # flash('You do not have permissions to edit this post.')
+        return redirect(url_for('get_all_posts'))
+
     edit_form = CreatePostForm(
         title=post.title,
         subtitle=post.subtitle,
@@ -164,11 +237,17 @@ def edit_post(post_id):
 
 
 @app.route("/delete/<int:post_id>")
+@login_required
 def delete_post(post_id):
 
     post_to_delete = db.session.query(BlogPost).get(post_id)
-    db.session.delete(post_to_delete)
-    db.session.commit()
+
+    if check_permissions(post_to_delete):
+        db.session.delete(post_to_delete)
+        db.session.commit()
+    # else:
+    #     This user lacks permissions to delete the post.
+    #     flash('You cannot perform this action.')
 
     return redirect(url_for('get_all_posts'))
